was able to implement clock with socket.io
