+++ index.js
@@ -1,60 +1,30 @@
-const express = require('express');
-const allRoutes = require('./controllers');
-
-const sequelize = require('./config');
-
-// allow cross-origin resource sharing (CORS)
-const cors = require("cors")
-
-// Sets up the Express App
-const app = express();
-const PORT = process.env.PORT || 3000;
-// =============================================================
-
-// set up socket.io
+const express = require('express')
+const cors = require('cors')
+const socketIo = require('socket.io')
 const http = require('http')
+const PORT = process.env.PORT || 3000
+const app = express()
+app.use(cors())
 const server = http.createServer(app)
-const { Server } = require("socket.io")
-const io = new Server({
+const io = socketIo(server,{ 
     cors: {
-      origin: `http://localhost:${PORT}`
+      origin: 'http://localhost:3001'
     }
-  })
+}) //in case server and client run on different urls
+io.on('connection',(socket)=>{
+  console.log('client connected: ',socket.id)
   
-
-console.log('===\n\n\ntest\n\n\n===')
-io.on('connection', (socket) => {
-    console.log('a user connected', socket.id);
-    socket.join('clock-room')
+  socket.join('clock-room')
   
   socket.on('disconnect',(reason)=>{
     console.log(reason)
   })
-  });
-
-  setInterval(()=>{
-    io.to('clock-room').emit('time', new Date())
+})
+setInterval(()=>{
+     io.to('clock-room').emit('time', new Date())
 },1000)
-console.log('===\n\n\ntest\n\n\n===')
-// =============================================================
 
-
-
-
-// Requiring our models for syncing
-const { User} = require('./models');
-
-
-// Sets up the Express app to handle data parsing
-app.use(express.urlencoded({ extended: true }));
-app.use(express.json());
-app.use(cors())
-
-app.use('/',allRoutes);
-
-
-sequelize.sync({ force: false }).then(function() {
-    app.listen(PORT, function() {
-    console.log('App listening on PORT ' + PORT);
-    });
-});
+server.listen(PORT, err=> {
+  if(err) console.log(err)
+  console.log('Server running on Port ', PORT)
+})
